fix(ci): green up all four CI platforms
- DateTime: compute the Unix-ns product in u64, not i64.
  DateTimeToUnixNs overflowed i64 for years > ~2262, and
  test_year_boundaries_sweep walks to 2550 -> SIGABRT under UBSan on the
  Linux sanitized job.

- TestRunner: author the AArch64 SetJmp/LongJmp as file-scope asm under
  GCC. GCC silently ignores __attribute__((naked)) on AArch64 (warns
  "'naked' attribute directive ignored") and emits a prologue that
  decrements SP, so the SP saved into the JmpBuf is wrong; LongJmp then
  restores it and every deadend test crashed in g_test_abort_jmp on the
  arm64 job. Clang honours naked (incl. macOS arm64), so that path is
  unchanged. GCC/AArch64 is Linux/ELF-only here, so no Mach-O underscore
  is needed. Verified red->green under qemu-aarch64.

- Socket: gate the IPv6 raw sockaddr byte check to PLATFORM_LINUX. macOS
  prefixes sin6_len and uses AF_INET6==30, Windows uses 23; the raw
  layout is Linux-specific (and mull, which the check feeds, runs on
  Linux). Fixes the macOS job.

- ArgParse: give the parser an output sink (ArgParse.out, default NULL =
  FileStderr()) so --help and error text is captured through the library's
  own File IO instead of OS-level stream redirection. The capture tests now
  point `out` at a FileOpenTemp file and read it back -- fully portable, so
  they compile and run identically on Linux, macOS and Windows. The old
  tests #include'd <unistd.h> and used pipe/dup2, which broke the Windows
  clang-cl/msvc build. No platform #if, no test gap across OSes.

+++ Source/Misra/Std/DateTime.c
@@ -74,7 +74,7 @@ u64 DateTimeToUnixNs(DateTime dt) {
     i64 days      = days_from_civil(dt.year, dt.month, dt.day);
     i64 local_sec = days * SECS_PER_DAY + (i64)dt.hour * 3600 + (i64)dt.minute * 60 + (i64)dt.second;
     i64 utc_sec   = local_sec - (i64)dt.utc_offset_seconds;
-    return (u64)(utc_sec * NS_PER_SEC + (i64)dt.nanosecond);
+    return (u64)utc_sec * (u64)NS_PER_SEC + (u64)dt.nanosecond;
 }
 
 i32 DateTimeCompare(DateTime a, DateTime b) {
